+++ src/cam-database/types.ts
@@ -1,29 +1,20 @@
-type Cam = {
+// Type Definitions
+type Brand = {
     id: string;
     name: string;
-    brand_id: string;
-    model_id: string;
-    size: {
-        inches: { min: number; max: number };
-        mm: { min: number; max: number };
-    };
-    color: string;
-    regex: RegExp;
 };
 type Model = {
     id: string;
     name: string;
-    cams: Cam[];
+    brand: Brand;
 };
-type Brand = {
+type Cam = {
     id: string;
     name: string;
-    models: Model[];
-};
-type Database = {
-    brands: Brand[];
-    models: Model[];
-    cams: Cam[];
+    size: { inches: [number, number]; mm: [number, number] };
+    color: string;
+    regex: RegExp;
+    model: Model;
 };
 
-export { Cam, Model, Brand, Database };
+export { Brand, Model, Cam };
